moved firepad components to editor folder
